+++ frontend/appflowy_flutter/integration_test/desktop/document/document_with_cover_image_test.dart
@@ -13,8 +13,15 @@ import '../../shared/emoji.dart';
 import '../../shared/util.dart';
 
 void main() {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-  RecentIcons.enable = false;
+  setUpAll(() {
+    IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+    RecentIcons.enable = false;
+  });
+
+  tearDownAll(() {
+    RecentIcons.enable = true;
+  });
+
   group('cover image:', () {
     testWidgets('document cover tests', (tester) async {
       await tester.initializeAppFlowy();
