+++ frontend/appflowy_flutter/lib/shared/icon_emoji_picker/recent_icons.dart
@@ -13,51 +13,16 @@ class RecentIcons {
   static final Map<String, List<String>> _dataMap = {};
   static bool _loaded = false;
   static const maxLength = 20;
+
+  /// To prevent the Recent Icon feature from affecting the unit tests of the Icon Selector.
   @visibleForTesting
   static bool enable = true;
 
-  static Future _save() async {
-    await getIt<KeyValueStorage>().set(
-      KVKeys.kRecentIcons,
-      jsonEncode(_dataMap),
-    );
-  }
-
-  static Future _load() async {
-    if (_loaded || !enable) return;
-    final v = await getIt<KeyValueStorage>().get(KVKeys.kRecentIcons);
-    try {
-      final data = jsonDecode(v ?? '') as Map;
-      _dataMap.clear();
-      _dataMap.addAll(
-        data.map(
-          (k, v) => MapEntry<String, List<String>>(k, List<String>.from(v)),
-        ),
-      );
-    } on FormatException catch (e) {
-      Log.error('RecentIcons load with :$v', e);
-    } on TypeError catch (e) {
-      Log.error('RecentIcons load with :$v', e);
-    }
-    _loaded = true;
-  }
-
-  static Future _put(FlowyIconType key, String value) async {
-    await _load();
-    if (!enable) return;
-    final list = _dataMap[key.name] ?? [];
-    list.remove(value);
-    list.insert(0, value);
-    if (list.length > maxLength) list.removeLast();
-    _dataMap[key.name] = list;
-    await _save();
-  }
-
-  static Future putEmoji(String id) async {
+  static Future<void> putEmoji(String id) async {
     await _put(FlowyIconType.emoji, id);
   }
 
-  static Future putIcon(Icon icon) async {
+  static Future<void> putIcon(Icon icon) async {
     await _put(
       FlowyIconType.icon,
       jsonEncode(
@@ -79,11 +44,52 @@ class RecentIcons {
       return iconList
           .map((e) => Icon.fromJson(jsonDecode(e) as Map<String, dynamic>))
           .toList();
-    } on FormatException catch (e) {
-      Log.error('RecentIcons getIcons with :$iconList', e);
-    } on TypeError catch (e) {
+    } catch (e) {
       Log.error('RecentIcons getIcons with :$iconList', e);
     }
     return [];
+  }
+
+  static Future<void> _save() async {
+    await getIt<KeyValueStorage>().set(
+      KVKeys.recentIcons,
+      jsonEncode(_dataMap),
+    );
+  }
+
+  static Future<void> _load() async {
+    if (_loaded || !enable) {
+      return;
+    }
+    final storage = getIt<KeyValueStorage>();
+    final value = await storage.get(KVKeys.recentIcons);
+    if (value == null || value.isEmpty) {
+      _loaded = true;
+      return;
+    }
+    try {
+      final data = jsonDecode(value) as Map;
+      _dataMap
+        ..clear()
+        ..addAll(
+          Map<String, List<String>>.from(
+            data.map((k, v) => MapEntry(k, List<String>.from(v))),
+          ),
+        );
+    } catch (e) {
+      Log.error('RecentIcons load failed with: $value', e);
+    }
+    _loaded = true;
+  }
+
+  static Future<void> _put(FlowyIconType key, String value) async {
+    await _load();
+    if (!enable) return;
+    final list = _dataMap[key.name] ?? [];
+    list.remove(value);
+    list.insert(0, value);
+    if (list.length > maxLength) list.removeLast();
+    _dataMap[key.name] = list;
+    await _save();
   }
 }
